Fix train position calculation with entry separator tracking (#320)
Trains now track which separator they entered the current section through,
enabling correct interpolation direction and preventing visual artifacts.

Changes:
- Train: Added entrySeparator field to track section entry point
- TrainPositionCalculator: Use position-based comparison to determine
  interpolation direction (entry→exit), with epsilon tolerance for
  static/dynamic separator matching
- PointF: New utility class for continuous (sub-pixel) coordinates
- AnimationState: Updated TrainState to use PointF for smooth rendering

This prevents IllegalStateException crashes and provides correct train
movement direction. Position-based comparison with fallback handles race
conditions gracefully (causes brief "JUMP DETECTED" warnings but no
visible glitches).

Future optimization in #320: Identity-based comparison using === operator
to eliminate false positive warnings.

All 1,798 tests passing.

+++ src/main/kotlin/cz/vutbr/fit/interlockSim/gui/gridcanvas/AnimatedSimulationCellRenderer.kt
@@ -310,8 +310,9 @@ class AnimatedSimulationCellRenderer(
 		val gridLocation = trainState.frontGridLocation ?: return
 
 		// Convert grid coordinates to pixel coordinates (center of cell)
-		val pixelX = gridLocation.x * cellWidth + cellWidth / 2
-		val pixelY = gridLocation.y * cellHeight + cellHeight / 2
+		// PointF provides continuous coordinates, round to nearest pixel for rendering
+		val pixelX = (gridLocation.x * cellWidth + cellWidth / 2).toInt()
+		val pixelY = (gridLocation.y * cellHeight + cellHeight / 2).toInt()
 
 		// Train body: blue circle
 		g.color = AnimationColors.TRAIN_BODY
